rtc: hctosys: Ensure system time doesn't overflow time_t
Issue: LIN10-6409

1/2 [
Author: Alexandre Belloni
Email: alexandre.belloni@bootlin.com
Subject: rtc: hctosys: Ensure system time doesn't overflow time_t
Date: Thu, 8 Mar 2018 23:27:31 +0100

commit b3a5ac42ab18b7d1a8f2f072ca0ee76a3b754a43 upstream

On 32bit platforms, time_t is still a signed 32bit long. If it is
overflowed, userspace and the kernel cant agree on the current system time.
This causes multiple issues, in particular with systemd:
https://github.com/systemd/systemd/issues/1143

A good workaround is to simply avoid using hctosys which is something I
greatly encourage as the time is better set by userspace.

However, many distribution enable it and use systemd which is rendering the
system unusable in case the RTC holds a date after 2038 (and more so after
2106). Many drivers have workaround for this case and they should be
eliminated so there is only one place left to fix when userspace is able to
cope with dates after the 31bit overflow.

Acked-by: Arnd Bergmann <arnd@arndb.de>
Signed-off-by: Alexandre Belloni <alexandre.belloni@bootlin.com>
Signed-off-by: Zhang Xiao <xiao.zhang@windriver.com>
]

2/2 [
Author: Maciej W. Rozycki
Email: macro@linux-mips.org
Subject: rtc: hctosys: Add missing range error reporting
Date: Mon, 5 Nov 2018 03:48:25 +0000

commit 7ce9a992ffde8ce93d5ae5767362a5c7389ae895 upstream

Fix an issue with the 32-bit range error path in `rtc_hctosys' where no
error code is set and consequently the successful preceding call result
from `rtc_read_time' is propagated to `rtc_hctosys_ret'.  This in turn
makes any subsequent call to `hctosys_show' incorrectly report in sysfs
that the system time has been set from this RTC while it has not.

Set the error to ERANGE then if we can't express the result due to an
overflow.

Signed-off-by: Maciej W. Rozycki <macro@linux-mips.org>
Fixes: b3a5ac42ab18 ("rtc: hctosys: Ensure system time doesn't overflow time_t")
Cc: stable@vger.kernel.org # 4.17+
Signed-off-by: Alexandre Belloni <alexandre.belloni@bootlin.com>
Signed-off-by: Zhang Xiao <xiao.zhang@windriver.com>
]
Signed-off-by: Yue Tao <Yue.Tao@windriver.com>