RDMA/cxgb4: Do not dma memory off of the stack
Issue: LIN10-6527

1/1 [
Author: Li Zhou
Email: li.zhou@windriver.com
Subject: RDMA/cxgb4: Do not dma memory off of the stack
Date: Tue, 15 Oct 2019 15:42:28 +0800

commit 3840c5b78803b2b6cc1ff820100a74a092c40cbb upstream

Nicolas pointed out that the cxgb4 driver is doing dma off of the stack,
which is generally considered a very bad thing.  On some architectures it
could be a security problem, but odds are none of them actually run this
driver, so it's just a "normal" bug.

Resolve this by allocating the memory for a message off of the heap
instead of the stack.  kmalloc() always will give us a proper memory
location that DMA will work correctly from.

Link: https://lore.kernel.org/r/20191001165611.GA3542072@kroah.com
Reported-by: Nicolas Waisman <nico@semmle.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Tested-by: Potnuri Bharat Teja <bharat@chelsio.com>
Signed-off-by: Jason Gunthorpe <jgg@mellanox.com>
[lz: adapted the patch to solve context issues]
Signed-off-by: Li Zhou <li.zhou@windriver.com>
]
Signed-off-by: Yue Tao <Yue.Tao@windriver.com>