HID: Fix assumption that devices have inputs
Issue: LIN10-6787

1/1 [
Author: Alan Stern
Email: stern@rowland.harvard.edu
Subject: HID: Fix assumption that devices have inputs
Date: Thu, 3 Oct 2019 14:53:59 -0400

commit d9d4b1e46d9543a82c23f6df03f4ad697dab361b upstream

The syzbot fuzzer found a slab-out-of-bounds write bug in the hid-gaff
driver.  The problem is caused by the driver's assumption that the
device must have an input report.  While this will be true for all
normal HID input devices, a suitably malicious device can violate the
assumption.

The same assumption is present in over a dozen other HID drivers.
This patch fixes them by checking that the list of hid_inputs for the
hid_device is nonempty before allowing it to be used.

Reported-and-tested-by: syzbot+403741a091bf41d4ae79@syzkaller.appspotmail.com
Signed-off-by: Alan Stern <stern@rowland.harvard.edu>
CC: <stable@vger.kernel.org>
Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>
Signed-off-by: Yue Tao <Yue.Tao@windriver.com>
]
Signed-off-by: Yue Tao <Yue.Tao@windriver.com>