NFC: pn533: fix use-after-free and memleaks
Issue: LIN10-6793

1/2 [
Author: Greg Kroah-Hartman
Email: gregkh@linuxfoundation.org
Subject: NFC: pn533: don't send USB data off of the stack
Date: Sun, 20 May 2018 15:19:46 +0200

commit dbafc28955fa6779dc23d1607a0fee5e509a278b upstream

It's amazing that this driver ever worked, but now that x86 doesn't
allow USB data to be sent off of the stack, it really does not work at
all.  Fix this up by properly allocating the data for the small
"commands" that get sent to the device off of the stack.

We do this for one command by having a whole urb just for ack messages,
as they can be submitted in interrupt context, so we can not use
usb_bulk_msg().  But the poweron command can sleep (and does), so use
usb_bulk_msg() for that transfer.

Reported-by: Carlos Manuel Santos <cmmpsantos@gmail.com>
Cc: Samuel Ortiz <sameo@linux.intel.com>
Cc: Stephen Hemminger <stephen@networkplumber.org>
Cc: stable <stable@vger.kernel.org>
Reviewed-by: Johan Hovold <johan@kernel.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Signed-off-by: Zhixiong Chi <zhixiong.chi@windriver.com>
]

2/2 [
Author: Johan Hovold
Email: johan@kernel.org
Subject: NFC: pn533: fix use-after-free and memleaks
Date: Mon, 7 Oct 2019 18:40:59 +0200

commit 6af3aa57a0984e061f61308fe181a9a12359fecc upstream

The driver would fail to deregister and its class device and free
related resources on late probe errors.

Reported-by: syzbot+cb035c75c03dbe34b796@syzkaller.appspotmail.com
Fixes: 32ecc75ded72 ("NFC: pn533: change order operations in dev registation")
Signed-off-by: Johan Hovold <johan@kernel.org>
Signed-off-by: Jakub Kicinski <jakub.kicinski@netronome.com>
Signed-off-by: Zhixiong Chi <zhixiong.chi@windriver.com>
]
Signed-off-by: Yue Tao <Yue.Tao@windriver.com>