ext4: add more paranoia checking in ext4_expand_extra_isize handling
Issue: LIN1018-5469

1/1 [
Author: Theodore Ts'o
Email: tytso@mit.edu
Subject: ext4: add more paranoia checking in ext4_expand_extra_isize handling
Date: Thu, 7 Nov 2019 21:43:41 -0500

commit 4ea99936a1630f51fc3a2d61a58ec4a1c4b7d55a upstream

It's possible to specify a non-zero s_want_extra_isize via debugging
option, and this can cause bad things(tm) to happen when using a file
system with an inode size of 128 bytes.

Add better checking when the file system is mounted, as well as when
we are actually doing the trying to do the inode expansion.

Link: https://lore.kernel.org/r/20191110121510.GH23325@mit.edu
Reported-by: syzbot+f8d6f8386ceacdbfff57@syzkaller.appspotmail.com
Reported-by: syzbot+33d7ea72e47de3bdf4e1@syzkaller.appspotmail.com
Reported-by: syzbot+44b6763edfc17144296f@syzkaller.appspotmail.com
Signed-off-by: Theodore Ts'o <tytso@mit.edu>
Cc: stable@kernel.org
Signed-off-by: Stefan Ghinea <stefan.ghinea@windriver.com>
]
Signed-off-by: Yue Tao <Yue.Tao@windriver.com>