marvell-cn96xx: Enable the SCMI support
Issue: LINUXEXEC-6287

ARM System Control and Management Interface (SCMI) protocol is a set
of operating system-independent software interfaces that are used in
system management. The octeontx2 platform firmware does have the SCMI
support, so enable the corresponding kernel drivers. The SCMI has a
dependency on the mailbox, so also enable it.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>