nxp-imx6: enable firmware load related kernel config for BSP nxp-imx6
Issue: LINUXEXEC-2128

There is a commit 95e2acb8f7c7("firmware_loader: Enable
FW_LOADER_USER_HELPER") to enable firmware loading related
config as default. It is not resaonable, so remove this
patch and enable these config in cfg file
In addtion, some drivers(for example, sdma, vpu and so on) need
to load firmware, so these config should be enabled.

Signed-off-by: Meng Li <Meng.Li@windriver.com>