xsk: Add missing check on user supplied headroom size
Issue: LIN1018-6095

1/1 [
Author: Magnus Karlsson
Email: magnus.karlsson@intel.com
Subject: xsk: Add missing check on user supplied headroom size
Date: Tue, 14 Apr 2020 09:35:15 +0200

commit ad8fb61c184fe0f8d1e0b5b954d010fb9f94a6ee from linux-stable
[ Upstream commit 99e3a236dd43d06c65af0a2ef9cb44306aef6e02 ]

Add a check that the headroom cannot be larger than the available
space in the chunk. In the current code, a malicious user can set the
headroom to a value larger than the chunk size minus the fixed XDP
headroom. That way packets with a length larger than the supported
size in the umem could get accepted and result in an out-of-bounds
write.

Fixes: c0c77d8fb787 ("xsk: add user memory registration support sockopt")
Reported-by: Bui Quang Minh <minhquangbui99@gmail.com>
Signed-off-by: Magnus Karlsson <magnus.karlsson@intel.com>
Signed-off-by: Daniel Borkmann <daniel@iogearbox.net>
Link: https://bugzilla.kernel.org/show_bug.cgi?id=207225
Link: https://lore.kernel.org/bpf/1586849715-23490-1-git-send-email-magnus.karlsson@intel.com
Signed-off-by: Sasha Levin <sashal@kernel.org>
[SG: backport of ad8fb61 from linux-4.19.y branch of linux-stable]
Signed-off-by: Stefan Ghinea <stefan.ghinea@windriver.com>
]
Signed-off-by: Yue Tao <Yue.Tao@windriver.com>