powerpc: Fix assembler warning about r0
Issue: LIN1018-5909

1/1 [
Author: Li Wang
Email: li.wang@windriver.com
Subject: powerpc: Fix assembler warning about r0
Date: Tue, 11 Aug 2020 10:49:19 +0800

The assembler says:
  arch/powerpc/platforms/85xx/sleep.S:1174: Warning: invalid register expression

It's objecting to the use of r0 as the RA argument. That's because
when RA = 0 the literal value 0 is used, rather than the content of
r0, making the use of r0 in the source potentially confusing.

Fix it to use a literal 0, the generated code is identical.

Signed-off-by: Li Wang <li.wang@windriver.com>
]
Signed-off-by: Yue Tao <Yue.Tao@windriver.com>