nfsd: apply umask on fs without ACL support
Issue: LIN10-7669

1/1 [
Author: J. Bruce Fields
Email: bfields@redhat.com
Subject: nfsd: apply umask on fs without ACL support
Date: Tue, 16 Jun 2020 16:43:18 -0400

commit 22cf8419f1319ff87ec759d0ebdff4cbafaee832 upstream.

The server is failing to apply the umask when creating new objects on
filesystems without ACL support.

To reproduce this, you need to use NFSv4.2 and a client and server
recent enough to support umask, and you need to export a filesystem that
lacks ACL support (for example, ext4 with the "noacl" mount option).

Filesystems with ACL support are expected to take care of the umask
themselves (usually by calling posix_acl_create).

For filesystems without ACL support, this is up to the caller of
vfs_create(), vfs_mknod(), or vfs_mkdir().

Reported-by: Elliott Mitchell <ehem+debian@m5p.com>
Reported-by: Salvatore Bonaccorso <carnil@debian.org>
Tested-by: Salvatore Bonaccorso <carnil@debian.org>
Fixes: 47057abde515 ("nfsd: add support for the umask attribute")
Cc: stable@vger.kernel.org
Signed-off-by: J. Bruce Fields <bfields@redhat.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
[Backport from 4.19.y branch]
Signed-off-by: Zhixiong Chi <zhixiong.chi@windriver.com>
]
Signed-off-by: Yue Tao <Yue.Tao@windriver.com>