xfs: fix boundary test in xfs_attr_shortform_verify
Issue: LIN1018-6705

1/1 [
Author: Eric Sandeen
Email: sandeen@redhat.com
Subject: xfs: fix boundary test in xfs_attr_shortform_verify
Date: Wed, 26 Aug 2020 14:11:58 -0700

commit f4020438fab05364018c91f7e02ebdd192085933 upstream

The boundary test for the fixed-offset parts of xfs_attr_sf_entry in
xfs_attr_shortform_verify is off by one, because the variable array
at the end is defined as nameval[1] not nameval[].
Hence we need to subtract 1 from the calculation.

This can be shown by:

and verifications will fail when it's written to disk.

This only matters for a last attribute which has a single-byte name
and no value, otherwise the combination of namelen & valuelen will
push endp further out and this test won't fail.

Fixes: 1e1bbd8e7ee06 ("xfs: create structure verifier function for shortform xattrs")
Signed-off-by: Eric Sandeen <sandeen@redhat.com>
Reviewed-by: Darrick J. Wong <darrick.wong@oracle.com>
Signed-off-by: Darrick J. Wong <darrick.wong@oracle.com>
Reviewed-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: Liu Haitao <haitao.liu@windriver.com>
]
Signed-off-by: Yue Tao <Yue.Tao@windriver.com>