nxp-ls20xx: delete CONFIG_IOMMU_DEFAULT_PASSTHROUGH for LTS19
Issue: LIN1019-5597

The config CONFIG_IOMMU_DEFAULT_PASSTHROUGH will cause
VFIO show warning calltrace when DPDK runs on DPAA2.
It is originally for usb device, at that time, usb device
can not be identified if this config not set. But now,
the usb issue is fixed by another patch, so delete it.

Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>